Unify PatternDataFormat and PatternTimeFormat duplicated code
